Deep link and make photo module common
