+++ ClientApp/src/app/app.component.html
@@ -1,5 +1,5 @@
 <div id="boot" class="modal" *ngIf="!this.bootStatus">
-<div id="errorWindow" class="alert picasso-mac content noscroll text-center" style="width: 20%;" *ngIf="!this.bootStatus" >
+<div id="errorWindow" class="alert picasso-mac content noscroll text-center" *ngIf="!this.bootStatus" >
   <div class="inner">
       <p style="font-weight: normal;">Welcome to Macintosh.</p>
       <br />
