+++ ClientApp/src/app/photo-upload/photo-upload.component.html
@@ -19,23 +19,10 @@
       <img src="../../assets/img/quicktake.png" alt="Photos" />
     </div>
 <h3>{{this.authenticateStatus}}</h3>
-<div>
-  <input #fileInput class="form-control" type="file" accept="image/x-png,image/jpeg,image/heic,image/heif"
-      multiple (change)="onChange($event)">
-
-  <button (click)="onUpload(); fileInput.value=''">
-    Upload your files
-  </button>
-<!-- list of files-->
-<div *ngFor="let p of this.fileLoader; let i = index">
-  <p *ngIf="p.uploadMessage === ''">{{p.file.name}}</p>
-  <p *ngIf="p.uploadMessage != ''">{{p.uploadMessage}}</p>
-  <p><ngb-progressbar type="info" height="20px" [value]=p.uploadPercent></ngb-progressbar></p>
-  <hr>
+<div *ngIf="this.authenticated">
+  <app-upload-photo-common>
+  </app-upload-photo-common>
 </div>
-
-</div>
-
 </cdk-virtual-scroll-viewport>
 </div>
 </div>
