+++ ClientApp/src/app/services/upload-photo-common/upload-photo-common.component.html
@@ -1,1 +1,16 @@
-<p>upload-photo-common works!</p>
+<div>
+  <input #fileInput class="form-control" type="file" accept="image/x-png,image/jpeg,image/heic,image/heif"
+      multiple (change)="onChange($event)">
+
+  <button (click)="onUpload(); fileInput.value=''">
+    Upload your files
+  </button>
+<!-- list of files-->
+<div *ngFor="let p of this.fileLoader; let i = index">
+  <p *ngIf="p.uploadMessage === ''">{{p.file.name}}</p>
+  <p *ngIf="p.uploadMessage != ''">{{p.uploadMessage}}</p>
+  <p><ngb-progressbar type="info" height="20px" [value]=p.uploadPercent></ngb-progressbar></p>
+  <hr>
+</div>
+
+</div>
