+++ ClientApp/src/app/wedding/photos/photos.component.ts
@@ -1,11 +1,10 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { DomSanitizer } from '@angular/platform-browser';
-import { BehaviorSubject, Observable, Subject, take } from 'rxjs';
+import { BehaviorSubject, Observable, Subject, take, takeUntil } from 'rxjs';
 import { photoListDto } from 'src/app/data/data';
 import { AuthService } from 'src/app/services/auth.service';
 import { DataService } from 'src/app/services/data.service';
-import { Guid } from 'typescript-guid';
 import { NgbCarouselConfig } from '@ng-bootstrap/ng-bootstrap';
+import { EventService } from 'src/app/services/event.service';
 
 
 @Component({
@@ -16,7 +15,6 @@ import { NgbCarouselConfig } from '@ng-bootstrap/ng-bootstrap';
 export class PhotosComponent implements OnInit, OnDestroy{
 
   destroy$: Subject<boolean> = new Subject<boolean>();
-  files!: File[]; // array of files for future use
   message: string | undefined;
   progress: number | undefined;
   galleryDisplay: boolean = false;
@@ -24,10 +22,12 @@ export class PhotosComponent implements OnInit, OnDestroy{
   storageUri: string = "https://photo.kevinandaustin.com/";
   sasKey: string = "";
 
-  constructor (private dataService: DataService, private sanitizer: DomSanitizer, private authService: AuthService) { }
+  constructor (private dataService: DataService, private eventService: EventService,
+    private authService: AuthService) { }
 
   windowVisible: boolean = true;
   galleryExpanded: boolean = true;
+  activeSlide: string = "slide-0";
 
   private photoListSubject : BehaviorSubject<photoListDto[]> | undefined;
   public photoList$ : Observable<photoListDto[]> | undefined;
@@ -39,8 +39,11 @@ export class PhotosComponent implements OnInit, OnDestroy{
 
     this.sasKey = this.authService.getSasToken();
 
-    this.getThumbsFromApi();
+    this.eventService.emptyTrash.subscribe(() => {
+      this.getThumbsFromApi();
+    });
 
+    this.getThumbsFromApi();
   }
 
   toggleCollapse(winType: string): void {
@@ -54,44 +57,16 @@ export class PhotosComponent implements OnInit, OnDestroy{
     }
   }
 
-  async onUpload(): Promise<void> {
-    // basic filter, not 100% reliable
-
-    for(let index = 0; index < this.files.length; index++)
-    {
-      var newfile = this.files[index];
-      console.log(newfile);
-      if (newfile.type.startsWith("image/"))
-      {
-        // Create form data
-        const formData = new FormData();
-        formData.append("file", newfile, newfile.name);
-
-        this.dataService.savePhotoFile(formData).subscribe(x => {
-          console.log(x);
-          this.message = "Your file has been uploaded!";
-          this.getThumbsFromApi();
-        });
-      }
-    };
-  }
-
   getThumbsFromApi()
   {
-    this.dataService.getThumbnails().subscribe(pr => {
+    this.dataService.getThumbnails().pipe(takeUntil(this.destroy$)).subscribe(pr => {
       this.photoListSubject!.next(pr)
     });
   }
 
-  onChange(event: any) {
-    this.files = event.target.files;
-    this.message = "";
-  }
-
   showGallery(photoIndex: number) {
+    this.activeSlide = "slide-" + photoIndex.toFixed(0).toString();
     this.galleryDisplay = true;
-    console.log('showGallery' + this.galleryDisplay + ' ' + photoIndex);
-    // future use, filter to get the gallery pic and set as active in win
   }
 
   ngOnDestroy() {
